Moved bin logic to specific package to export the internal logic
